Tidy up imports and move similarity API to recommend
