+++ server/recommend/routes.py
@@ -53,3 +53,54 @@ async def negative_update(request: Request, term: str):
     from server.users.distance_utils import negative_move_session_vector
 
     return await update(request, term, negative_move_session_vector)
+
+
+@recommend_blueprint.route('/similarity/<term>', methods=['GET'])
+async def similarity(request: Request, term: str):
+    """
+    Measure how likely the current user is to be interested in the specified term.
+    :param request:
+    :param term:
+    :return: 200 OK with similarity score if user exists. 404 NOT_FOUND if the user doesn't exist.
+    """
+    from server.users.user import User
+
+    user_id = get_user_id(request)
+    if user_id is not None:
+        return await similarity(request, user_id, term)
+    raise InvalidUsage("Must supply '%s' cookie" % User.user_id_key)
+
+
+@recommend_blueprint.route('/similarity/<user_id>/<term>', methods=['GET'])
+async def similarity(request: Request, user_id: str, term: str):
+    """
+    Measure how likely this user is to be interested in the specified term
+    :param request:
+    :param user_id:
+    :param term:
+    :return: 200 OK with similarity score if user exists. 404 NOT_FOUND if the user doesn't exist.
+    """
+    from server.users import get_user
+    from sanic.response import json
+
+    user = await get_user(user_id)
+
+    if user is not None:
+        from server.word_embedding.supervised_models import load_model, SupervisedModels
+        from server.word_embedding.utils import cosine_sim
+
+        model = load_model(SupervisedModels.ONS)
+
+        term_vector = model.get_sentence_vector(term)
+        user_vector = await user.get_user_vector()
+
+        sim = cosine_sim(user_vector, term_vector)
+
+        response = {
+            'user_id': user_id,
+            'term': term,
+            'similarity': sim
+        }
+
+        return json(response, 200)
+    return json("User '%s' not found" % user_id, 404)
